+++ components/list/index.js
@@ -1,6 +1,8 @@
-import List from './List'
+import { List, Group, Divider } from './List'
 import Item from './Item'
 
 List.Item = Item
+List.Group = Group
+List.Divider = Divider
 
 export default List