+++ kitchen-sink/demos/Loadings.js
@@ -9,13 +9,27 @@ export default {
   view: (state, actions) => {
     return (
       <ContentBlock title="Loading">
-        <p onclick={e => {
-          Loading.show()
-          setTimeout(() => {
-            Loading.hide()
-          }, 2000)
-        }}
-        ><a>loading</a></p>
+        <p
+          onclick={e => {
+            Loading.show()
+            setTimeout(() => {
+              Loading.hide()
+            }, 2000)
+          }}
+        >
+          <a>loading</a>
+        </p>
+
+        {/* <p
+          onclick={e => {
+            Loading.show('loading...')
+            // setTimeout(() => {
+            //   Loading.hide()
+            // }, 2000)
+          }}
+        >
+          <a>loading text</a>
+        </p> */}
       </ContentBlock>
     )
   }